refactor: extract button to separate component
