+++ packages/shared/src/features/opportunity/components/UploadCVButton.tsx
@@ -1,0 +1,58 @@
+import React, { useCallback } from 'react';
+import type { ReactElement } from 'react';
+import { useQuery } from '@tanstack/react-query';
+import { fileValidation, useUploadCv } from '../../profile/hooks/useUploadCv';
+import { UploadButton } from '../../../components/buttons/UploadButton';
+import { ButtonSize, ButtonVariant } from '../../../components/buttons/common';
+import { useAuthContext } from '../../../contexts/AuthContext';
+import { useToastNotification } from '../../../hooks';
+import { useUpdateQuery } from '../../../hooks/useUpdateQuery';
+import { getCandidatePreferencesOptions } from '../queries';
+
+export const UploadCVButton = (): ReactElement => {
+  const { user } = useAuthContext();
+  const { displayToast } = useToastNotification();
+
+  const opts = getCandidatePreferencesOptions(user?.id);
+  const updateQuery = useUpdateQuery(opts);
+
+  const { data: preferences } = useQuery(opts);
+
+  const { onUpload: uploadCv, isPending: isUploadCvPending } = useUploadCv({
+    shouldOpenModal: false,
+    onUploadSuccess: () => displayToast('CV uploaded successfully!'),
+  });
+
+  const handleCVUpload = useCallback(
+    async (files: File[]) => {
+      if (files.length === 0) {
+        return;
+      }
+      const [, set] = updateQuery;
+      await uploadCv(files[0]);
+
+      set({
+        ...preferences,
+        cv: {
+          blob: user.id,
+          contentType: files[0].type,
+          lastModified: new Date(),
+        },
+      });
+    },
+    [preferences, updateQuery, uploadCv, user.id],
+  );
+
+  return (
+    <UploadButton
+      className="mb-4 mr-auto"
+      size={ButtonSize.Small}
+      variant={ButtonVariant.Subtle}
+      validation={fileValidation}
+      loading={isUploadCvPending}
+      onFilesDrop={handleCVUpload}
+    >
+      Upload PDF
+    </UploadButton>
+  );
+};
