+++ packages/webapp/pages/settings/job-preferences.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useState } from 'react';
+import React, { useEffect, useState } from 'react';
 import type { ReactElement } from 'react';
 import type { NextSeoProps } from 'next-seo';
 import {
@@ -44,11 +44,7 @@ import {
 } from '@dailydotdev/shared/src/hooks';
 import { ActionType } from '@dailydotdev/shared/src/graphql/actions';
 import { Tooltip } from '@dailydotdev/shared/src/components/tooltip/Tooltip';
-import {
-  fileValidation,
-  useUploadCv,
-} from '@dailydotdev/shared/src/features/profile/hooks/useUploadCv';
-import { UploadButton } from '@dailydotdev/shared/src/components/buttons/UploadButton';
+import { UploadCVButton } from '@dailydotdev/shared/src/features/opportunity/components/UploadCVButton';
 import { getSettingsLayout } from '../../components/layouts/SettingsLayout';
 import { defaultSeo } from '../../next-seo';
 import { getTemplatedTitle } from '../../components/layouts/utils';
@@ -116,31 +112,6 @@ const JobPreferencesPage = (): ReactElement => {
       displayToast('Failed to remove uploaded CV. Please try again.');
     },
   });
-
-  const { onUpload: uploadCv, isPending: isUploadCvPending } = useUploadCv({
-    shouldOpenModal: false,
-    onUploadSuccess: () => displayToast('CV uploaded successfully!'),
-  });
-
-  const handleCVUpload = useCallback(
-    async (files: File[]) => {
-      if (files.length === 0) {
-        return;
-      }
-      const [, set] = updateQuery;
-      await uploadCv(files[0]);
-
-      set({
-        ...preferences,
-        cv: {
-          blob: user.id,
-          contentType: files[0].type,
-          lastModified: new Date(),
-        },
-      });
-    },
-    [preferences, updateQuery, uploadCv, user.id],
-  );
 
   const modeDisabled = preferences?.status === CandidateStatus.DISABLED;
 
@@ -287,16 +258,7 @@ const JobPreferencesPage = (): ReactElement => {
                 </Typography>
               )}
 
-              <UploadButton
-                className="mb-4 mr-auto"
-                size={ButtonSize.Small}
-                variant={ButtonVariant.Subtle}
-                validation={fileValidation}
-                loading={isUploadCvPending}
-                onFilesDrop={handleCVUpload}
-              >
-                Upload PDF
-              </UploadButton>
+              <UploadCVButton />
               <FeelingLazy />
             </FlexCol>
